Release/4.2

+++ src/main/java/stream/flarebot/flarebot/commands/moderation/PurgeCommand.java
@@ -1,19 +1,22 @@
 package stream.flarebot.flarebot.commands.moderation;
 
-import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.Member;
 import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.MessageEmbed;
 import net.dv8tion.jda.core.entities.MessageHistory;
 import net.dv8tion.jda.core.entities.TextChannel;
 import net.dv8tion.jda.core.entities.User;
-import stream.flarebot.flarebot.commands.Command;
-import stream.flarebot.flarebot.commands.CommandType;
+import net.dv8tion.jda.core.exceptions.ErrorResponseException;
+import net.dv8tion.jda.core.exceptions.PermissionException;
+import stream.flarebot.flarebot.commands.*;
+import stream.flarebot.flarebot.mod.modlog.ModlogEvent;
+import stream.flarebot.flarebot.mod.modlog.ModlogHandler;
 import stream.flarebot.flarebot.objects.GuildWrapper;
+import stream.flarebot.flarebot.permissions.PerGuildPermissions;
 import stream.flarebot.flarebot.util.GeneralUtils;
 import stream.flarebot.flarebot.util.MessageUtils;
 
-import java.awt.Color;
 import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.EnumSet;
@@ -30,18 +33,18 @@ public class PurgeCommand implements Command {
     @Override
     public void onCommand(User sender, GuildWrapper guild, TextChannel channel, Message message, String[] args, Member member) {
         if (args.length >= 1) {
-            User targetUser;
+            User targetUser = null;
             int amount;
-            targetUser = GeneralUtils.getUser(args[0], guild.getGuildId(), true);
-            if (targetUser == null) {
-                if (args[0].matches("\\d+")) {
-                    amount = GeneralUtils.getInt(args[0], -1);
-                } else {
+            if (args.length == 1 && args[0].matches("\\d+")) {
+                amount = GeneralUtils.getInt(args[0], -1);
+            } else if (args.length == 2 && args[1].matches("\\d+")) {
+                amount = GeneralUtils.getInt(args[1], -1);
+                try {
+                    targetUser = GeneralUtils.getUser(args[0], guild.getGuildId(), true);
+                } catch (ErrorResponseException e) {
                     MessageUtils.sendErrorMessage("That target user cannot be found, try mentioning them, using the user ID or using `all` to clear the entire chat.", channel);
                     return;
                 }
-            } else if (args.length == 2) {
-                amount = GeneralUtils.getInt(args[1], -1);
             } else {
                 MessageUtils.sendUsage(this, channel, sender, args);
                 return;
@@ -54,7 +57,7 @@ public class PurgeCommand implements Command {
             }
 
             // This will be a successful delete so limit here.
-            if (!guild.getPermissions().isCreator(sender)) {
+            if (!PerGuildPermissions.isCreator(sender)) {
                 long riotPolice = cooldowns.computeIfAbsent(channel.getGuild().getId(), n -> 0L);
                 if (System.currentTimeMillis() - riotPolice < cooldown) {
                     channel.sendMessage(MessageUtils.getEmbed(sender)
@@ -72,12 +75,11 @@ public class PurgeCommand implements Command {
                 return;
             }
             MessageHistory history = new MessageHistory(channel);
-            int toRetrieve = amount + 1;
+            int toRetrieve = amount;
             int i = 0;
+            message.delete().complete();
             outer:
-            while (toRetrieve > 0) { // I don't really know if this should be min...
-                // since deleting 10 of someone could be like 100 back yet this would request it 10 times.
-                // For now I will just request 100 here each time.
+            while (toRetrieve > 0) {
                 if (history.retrievePast((targetUser == null ? Math.min(toRetrieve, 100) : 100)).complete().isEmpty()) {
                     break;
                 }
@@ -85,28 +87,43 @@ public class PurgeCommand implements Command {
                 List<Message> toDelete = new ArrayList<>();
                 for (Message msg : history.getRetrievedHistory()) {
                     if (msg.getCreationTime().plusWeeks(2).isBefore(OffsetDateTime.now())) break outer;
+                    if (msg.getId().equals(message.getId())) continue;
                     if ((targetUser != null && msg.getAuthor().getId().equals(targetUser.getId())) || targetUser == null) {
                         toDelete.add(msg);
+                        // This is to fix stuff like purges being logged.
+                        //FlareBot.instance().getEvents().getRemovedByMeList().add(msg.getIdLong());
                         i++;
                         toRetrieve--;
                     }
                     if (toRetrieve == 0) break;
                 }
-                channel.deleteMessages(toDelete).complete();
+                try {
+                    if (toDelete.size() == 0) break;
+                    if (toDelete.size() == 1) {
+                        channel.deleteMessageById(toDelete.get(0).getId()).complete();
+                        break;
+                    }
+                    channel.deleteMessages(toDelete).complete();
+                } catch (PermissionException e) {
+                    MessageUtils.sendErrorMessage("There was a permissions error encountered when trying to do this!\n" + e.getMessage(), channel, sender);
+                    return;
+                } catch (ErrorResponseException e) {
+                    MessageUtils.sendErrorMessage("There was an unknown error when trying to do this!\n" + e.getMeaning(), channel, sender);
+                    return;
+                }
                 toDelete.clear();
             }
-            EmbedBuilder eb = new EmbedBuilder();
-            eb.setColor(Color.WHITE).setTitle(targetUser == null ? "Chat Purge" : "User Purge", null);
-            if (targetUser != null)
-                eb.addField("User", MessageUtils.getTag(targetUser) + " (" + targetUser.getId() + ")", true);
-            eb.addField("Responsible moderator", sender.getAsMention(), true);
-            eb.addField("Messages purged", String.valueOf((i - 1)), true);
-
-            guild.getAutoModConfig().postToModLog(eb.build());
-            MessageUtils.sendAutoDeletedMessage(MessageUtils.getEmbed(sender)
-                            .setDescription(String.format("Deleted `%s` messages!", i - 1)).build(),
-                    TimeUnit.SECONDS.toMillis(5), channel);
-        }
+            if (i > 0) {
+                ModlogHandler.getInstance().postToModlog(guild, ModlogEvent.FLAREBOT_PURGE, targetUser, sender, null,
+                        new MessageEmbed.Field("Messages purged", String.valueOf((i)), true));
+                MessageUtils.sendAutoDeletedMessage(MessageUtils.getEmbed(sender)
+                                .setDescription(String.format("Deleted `%s` messages!", i)).build(),
+                        TimeUnit.SECONDS.toMillis(5), channel);
+            } else {
+                MessageUtils.sendInfoMessage("We couldn't find any messages to purge!", channel, sender);
+            }
+        } else
+            MessageUtils.sendUsage(this, channel, sender, args);
     }
 
     @Override
@@ -133,6 +150,11 @@ public class PurgeCommand implements Command {
     @Override
     public String[] getAliases() {
         return new String[]{"clean"};
+    }
+
+    @Override
+    public boolean deleteMessage() {
+        return false;
     }
 
     @Override
